refactor: TickerResponseDto 수정
