+++ crypto-module/api-server/src/main/java/com/crypto_trader/api_server/presentation/HomeController.java
@@ -3,9 +3,7 @@ package com.crypto_trader.api_server.presentation;
 import com.crypto_trader.api_server.application.SimpleMarketService;
 import com.crypto_trader.api_server.application.TickerService;
 import com.crypto_trader.api_server.application.dto.CryptoDto;
-import com.crypto_trader.api_server.application.dto.TickerRequestDto;
-import com.crypto_trader.api_server.domain.Ticker;
-import org.springframework.beans.factory.annotation.Autowired;
+import com.crypto_trader.api_server.application.dto.TickerResponseDto;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -28,7 +26,7 @@ public class HomeController {
     }
 
     @GetMapping("/api/tickers")
-    public List<TickerRequestDto> getAllTickers() {
+    public List<TickerResponseDto> getAllTickers() {
         return tickerService.getTickers();
     }
 }
